Reduced storage SPDS1

+++ CHI_TECH/ChiMesh/MeshContinuum/chi_meshcontinuum.h
@@ -143,7 +143,7 @@ public:
   bool IsCellBndry(int cell_global_index = 0);
 
   void FindAssociatedVertices(chi_mesh::CellFace& cur_face,
-                              std::vector<int>& dof_mapping);
+                              std::vector<short>& dof_mapping);
 
   chi_mesh::Vector3 ComputeCentroidFromListOfNodes(const std::vector<int>& list);
